exist null일때 로직추가

+++ routes/postPage.js
@@ -667,20 +667,25 @@ router.delete("/delete", async (req, res) => {
     });
 
     const exist = verifyUser.dataValues.postImageUrl; // 현재 URL에 전달된 id값을 받아서 db찾음
-    const url = exist.split("/"); // exist 저장된 fileUrl을 가져옴
-    const delFileName = url[url.length - 1];
-    if (delFileName !== "A-fo_default.jpg") {
-      s3.deleteObject(
-        {
-          Bucket: "a-fo-bucket2",
-          Key: delFileName,
-        },
-        (err, data) => {
-          if (err) {
-            throw err;
+
+    if (exist) {
+      const url = exist.split("/"); // exist 저장된 fileUrl을 가져옴
+      const delFileName = url[url.length - 1];
+      if (delFileName !== "A-fo_default.jpg") {
+        s3.deleteObject(
+          {
+            Bucket: "a-fo-bucket2",
+            Key: delFileName,
+          },
+          (err, data) => {
+            if (err) {
+              throw err;
+            }
           }
-        }
-      );
+        );
+      }
+    } else {
+      console.log("이미지 없어서 패스~");
     }
 
     // 게시물이 있는 경우
